Following changes have been made,
* merge setup_aws.py into trigger_aws_test_run.py
* removed sleep
