+++ resources/setup_virtual_env.sh
@@ -41,7 +41,8 @@ virtualenv -p /usr/bin/python3.6 ./virtual_env
 install_requirement_txt(){
 print_message "installing requirements"
 pip install -r ./resources/requirements.txt
-sleep 60
+print_message "all requirements are installed"
+
 }
 
 switch_to_virtual_env(){
